+++ src/f32x4_.rs
@@ -327,6 +327,7 @@ impl_simd_float! {
     N = 4,
     Simd = f32x4,
     UnsignedT = u32,
+    UnsignedSimd = u32x4,
   }
   old_powf_simd_fn_name = pow_f32x4,