+++ src/f64x2_.rs
@@ -283,6 +283,7 @@ impl_simd_float! {
     N = 2,
     Simd = f64x2,
     UnsignedT = u64,
+    UnsignedSimd = u64x2,
   }
   old_powf_simd_fn_name = pow_f64x2,