+++ src/simd_uint.rs
@@ -322,6 +322,16 @@ macro_rules! impl_simd_uint {
       #[must_use]
       $fn_reduce_min
 
+      /// Returns the bit patterns of `self` reinterpreted as signed integers of
+      /// the same size.
+      #[inline]
+      #[must_use]
+      pub const fn cast_signed(self) -> $SignedSimd {
+        // SAFETY: Both types accept all bit-patterns and only contain
+        // initialized memory.
+        unsafe { core::mem::transmute::<$Simd, $SignedSimd>(self) }
+      }
+
       /// Shifts left each element of `self` by the corresponding element of
       /// `rhs`, without bounding `rhs`.
       ///